[Refactor] update type of time to string

+++ utils/time.go
@@ -5,10 +5,10 @@ import "time"
 /**
  * get Unix time that's expressed at millisecond
  */
-func Now() int32 {
+func Now() int64 {
 	return ParseToMillisecond(time.Now().UnixNano())
 }
 
-func ParseToMillisecond(target int64) int32 {
-	return int32(target / int64(time.Millisecond))
+func ParseToMillisecond(target int64) int64 {
+	return target / int64(time.Millisecond)
 }
